Move new metrics test utils to their own package
